+++ src/test/java/stepdefinition/DSalgoArraysStepDefinition.java
@@ -7,253 +7,6 @@ import io.cucumber.java.en.When;
 public class DSalgoArraysStepDefinition {
 
 	
-	@Given("The  user is on the {string} page")
-	public void the_user_is_on_the_page(String string) {
-	
-		
-	}
-	
-	@When("The user clicks a {string} link")
-	public void the_user_clicks_a_link(String string) {
-
-		
-	}
-	
-	@Then("The user should be redirected to {string} Page")
-	public void the_user_should_be_redirected_to_page(String string) {
-
-		
-	}
-
-	
-	@Given("The user is on the {string} Page")
-	public void the_user_is_on_the_page1(String string) {
-
-		
-		
-	}
-	
-	@When("The user clicks {string} button of {string} Page")
-	public void the_user_clicks_button_of_page(String string, String string2) {
-
-		
-		
-	}
-	
-	@Then("The user should be redirected to a page having an Python Editor")
-	public void the_user_should_be_redirected_to_a_page_having_an_python_editor() {
-
-		
-		
-	}
-
-	
-	@Given("The user is on the Python Editor")
-	public void the_user_is_on_the_python_editor() {
-
-		
-		
-	}
-	
-	@When("The user clicks the Run button without entering the code in the Editor")
-	public void the_user_clicks_the_run_button_without_entering_the_code_in_the_editor() {
-
-
-		
-	}
-	
-	@When("Click the Run button")
-	public void click_the_run_button() {
-
-		
-		
-	}
-	
-	
-	@Then("Nothing happens to the page and no error message should be displayed")
-	public void nothing_happens_to_the_page_and_no_error_message_should_be_displayed() {
-
-		
-		
-		
-	}
-
-
-
-	
-	@Given("The user is on the Python Editor")
-	public void the_user_is_on_the_python_editor1() {
-
-		
-	}
-	
-	@When("The user write the invalid code in Editor")
-	public void the_user_write_the_invalid_code_in_editor() {
-
-		
-		
-	}
-	
-	@When("Click the Run button")
-	public void click_the_run_button1() {
-
-		
-		
-	}
-	
-	
-	@Then("The user should able to see an error message in alert window")
-	public void the_user_should_able_to_see_an_error_message_in_alert_window() {
-
-		
-		
-	}
-
-
-	@Given("The user is on the Python Editor")
-	public void the_user_is_on_the_python_editor2() {
-
-		
-		
-	}
-	
-	
-	@When("The user write the valid code in Editor")
-	public void the_user_write_the_valid_code_in_editor() {
-
-		
-	}
-	
-	
-	@When("Click the Run button")
-	public void click_the_run_button2() {
-
-		
-		
-	}
-	
-	@Then("The user should able to see output in the console")
-	public void the_user_should_able_to_see_output_in_the_console() {
-
-		
-	}
-
-	
-	@Given("The  user is on the {string} page")
-	public void the_user_is_on_the_page3(String string) {
-
-		
-	}
-	
-	
-	@When("The user clicks {string} link")
-	public void the_user_clicks_link(String string) {
-
-		
-		
-	}
-	
-	
-	@Then("The user should be redirected to {string} page having links like {string} ,{string},{string} and {string}")
-	public void the_user_should_be_redirected_to_page_having_links_like_and(String string, String string2, String string3, String string4, String string5) {
-
-		
-		
-	}
-
-	@Given("The user is on the Python Editor")
-	public void the_user_is_on_the_python_editor4() {
-
-		
-	}
-	
-	
-	@When("The user clicks the Run button without entering the code in the Editor")
-	public void the_user_clicks_the_run_button_without_entering_the_code_in_the_editor1() {
-
-		
-	}
-	
-	
-	@Then("Nothing happens to the page and no error message is displayed")
-	public void nothing_happens_to_the_page_and_no_error_message_is_displayed1() {
-
-		
-		
-	}
-
-	@Given("The user is on the Python Editor")
-	public void the_user_is_on_the_python_editor5() {
-
-		
-		
-	}
-	
-	
-	@When("The user clicks the Run button without entering the code in the Editor")
-	public void the_user_clicks_the_run_button_without_entering_the_code_in_the_editor4() {
-
-		
-		
-	}
-	
-	
-	@Then("Nothing happens to the page and no error message is displayed")
-	public void nothing_happens_to_the_page_and_no_error_message_is_displayed() {
-
-		
-		
-		
-	}
-
-	
-	@Given("The user is on {string} page")
-	public void the_user_is_on_page(String string) {
-
-		
-	}
-	
-	
-	@When("The user clicks {string} link")
-	public void the_user_clicks_link1(String string) {
-
-		
-	}
-	
-	
-	@Then("The user should be redirected to {string} page contains a question,an tryEditor with Run and Submit buttons")
-	public void the_user_should_be_redirected_to_page_contains_a_question_an_try_editor_with_run_and_submit_buttons(String string) {
-
-		
-		
-	}
-
-	@Given("The user is on the Python Editor")
-	public void the_user_is_on_the_python_editor6() {
-	   
-	
-	}
-	
-	
-	@When("The user write the valid code in Editor")
-	public void the_user_write_the_valid_code_in_editor4() {
-
-		
-	}
-	
-	
-	@When("Click the Submit button")
-	public void click_the_submit_button() {
-
-	}
-
-	@Then("The user see an message {string}")
-	public void the_user_see_an_message(String string) {
-
-		
-		
-
-	}
