+++ src/test/java/stepdefinition/DSalgoDataStructuresStepDefinition.java
@@ -1,5 +1,134 @@
 package stepdefinition;
 
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
+
 public class DSalgoDataStructuresStepDefinition {
 
+	
+//The  user is on the selected data structure page
+	@Given("The  user is on the selected data structure page")
+	public void the_user_is_on_the_selected_data_structure_page() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	@When("The user clicks on Topics covered")
+	public void the_user_clicks_on_topics_covered() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	@Then("The user should be redirected to selected topic covered page")
+	public void the_user_should_be_redirected_to_selected_topic_covered_page() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	
+	
+	//Verify if user able to validate the python editor with empty data
+	
+	@Given("The user is on the Python Editor")
+	public void the_user_is_on_the_python_editor() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	@When("The user clicks the Run button without entering the code in the Editor")
+	public void the_user_clicks_the_run_button_without_entering_the_code_in_the_editor() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	@When("Click the Run button")
+	public void click_the_run_button() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	@Then("Nothing happens to the page and no error message should be displayed")
+	public void nothing_happens_to_the_page_and_no_error_message_should_be_displayed() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	
+	//Verify if user able to validate the python editor with valid data
+
+	@When("The user write the valid code in Editor")
+	public void the_user_write_the_valid_code_in_editor() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+
+	@Then("The user should able to see output in the console")
+	public void the_user_should_able_to_see_output_in_the_console() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	
+	
+	//The  user is on the "Time Complexity" page
+	
+	@Given("The  user is on the {string} page")
+	public void the_user_is_on_the_page(String string) {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	@When("The user clicks {string} link")
+	public void the_user_clicks_link(String string) {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	@Then("The user should be redirected to {string} page")
+	public void the_user_should_be_redirected_to_page(String string) {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	
+	//Verify if user able to access each data structure list
+	@Given("The user is on the Data Structures Home page")
+	public void the_user_is_on_the_data_structures_home_page() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	@When("The user clicks on each {string} button")
+	public void the_user_clicks_on_each_button(String string) {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	@Then("The user should be redirected to the selected data structure page")
+	public void the_user_should_be_redirected_to_the_selected_data_structure_page() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	
+	//Verify if user able to validate the python editor with invalid data
+
+	@When("The user write the invalid code in Editor")
+	public void the_user_write_the_invalid_code_in_editor() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+
+	@Then("The user should able to see an error message in alert window")
+	public void the_user_should_able_to_see_an_error_message_in_alert_window() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	
+	//The user is on the selected topic covered page
+	
+	@Given("The user is on the selected topic covered page")
+	public void the_user_is_on_the_selected_topic_covered_page() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	@When("The user clicks {string} button")
+	public void the_user_clicks_button(String string) {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	@Then("The user should be redirected to a page having an Python Editor")
+	public void the_user_should_be_redirected_to_a_page_having_an_python_editor() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+	
+	
 }
