+++ src/test/java/testrunner/TestRunner.java
@@ -7,7 +7,7 @@ import io.cucumber.testng.CucumberOptions;
 
 
 @CucumberOptions(plugin= {"pretty", "html:target/cucumber-reports/dsalgo.html"},tags="@sanity",
-features={"src/test/resources/features/Array.feature"},glue= "stepdefinition")
+features={"src/test/resources/features/DataStructure.feature"},glue= "stepdefinition")
 
 public class TestRunner extends AbstractTestNGCucumberTests {
