+++ test-output/old/Default suite/toc.html
@@ -22,7 +22,7 @@
 </tr></table>
 <table width='100%' class='test-failed'>
 <tr><td>
-<table style='width: 100%'><tr><td valign='top'>Default test (0/41/0)</td><td valign='top' align='right'>
+<table style='width: 100%'><tr><td valign='top'>Default test (0/7/0)</td><td valign='top' align='right'>
   <a href='Default test.html' target='mainFrame'>Results</a>
 </td></tr></table>
 </td></tr><p/>
